remove FirebaseDatabase, we will only use the FileStorageDatabase
